feat: 사용자가 업로드한 이미지인지 검증하는 기능 추가

+++ server/src/main/java/com/seb_main_004/whosbook/curation/service/CurationImageService.java
@@ -7,6 +7,8 @@ import com.seb_main_004.whosbook.curation.repository.CurationImageRepository;
 import com.seb_main_004.whosbook.exception.BusinessLogicException;
 import com.seb_main_004.whosbook.exception.ExceptionCode;
 import com.seb_main_004.whosbook.image.service.StorageService;
+import com.seb_main_004.whosbook.member.entity.Member;
+import com.seb_main_004.whosbook.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,17 +26,19 @@ import java.util.Optional;
 public class CurationImageService {
     private final CurationImageRepository curationImageRepository;
     private final StorageService storageService;
+    private final MemberService memberService;
     private final static String CURATION_IMAGE_PATH = "curationImages";
-
     // 이미지 등록
-    public CurationImage uploadCurationImage(MultipartFile image){
+    public CurationImage uploadCurationImage(MultipartFile image, String authenticatedEmail){
 
+        Member member = memberService.findVerifiedMemberByEmail(authenticatedEmail);
 
-        String key = storageService.makeObjectKey(image, CURATION_IMAGE_PATH);
-        String imagePath = storageService.store(image, CURATION_IMAGE_PATH);
+        String imageKey = storageService.makeObjectKey(image, CURATION_IMAGE_PATH, member.getMemberId());
+        String imagePath = storageService.store(image, imageKey);
 
-        CurationImage curationImage = new CurationImage(key, imagePath);
+        CurationImage curationImage = new CurationImage(imageKey, imagePath);
 
+        log.info("# 저장된 이미지 Owner ID : {}", getImageOwnerId(imageKey));
         log.info("# 이미지 업로드 성공! 업로드 된 이미지 URL : {}", curationImage.getPath());
 
         return curationImageRepository.save(curationImage);
@@ -56,21 +60,35 @@ public class CurationImageService {
        );
     }
 
-    public List<CurationImage> verifyCurationSaveImages(CurationPostDto postDto) {
+    public List<CurationImage> verifyCurationSaveImages(CurationPostDto postDto, long authenticatedMemberId) {
         List<Long> curationImageIds = postDto.getImageIds();
         String content = postDto.getContent();
 
         List<CurationImage> curationImages = new ArrayList<>();
+
         for (long curationImageId : curationImageIds){
+
             CurationImage curationImage = findVerifiedCurationImageById(curationImageId);
+            verifyImageOwner(curationImage, authenticatedMemberId);
+
             if (content.contains(curationImage.getImageKey())){
                curationImages.add(curationImage);
             } else {
                 deleteCurationImage(curationImageId);
             }
+
         }
 
         return curationImages;
+    }
+
+    private void verifyImageOwner(CurationImage curationImage, long authenticatedMemberId){
+        long imageOwnerId = Long.parseLong(getImageOwnerId(curationImage.getImageKey()));
+        if (imageOwnerId != authenticatedMemberId) throw new BusinessLogicException(ExceptionCode.IMAGE_CAN_NOT_SAVE);
+    }
+
+    private String getImageOwnerId(String imageKey) {
+        return imageKey.substring(imageKey.lastIndexOf("/") + 1, imageKey.lastIndexOf("_"));
     }
 
 }
